requirements.txt
psycopg2-binary
typesense
sentence-transformers
python-dotenv
